--- src/main/java/com/umc/mada/user/dto/UserResponseDto.java
@@ -1,5 +1,0 @@
-package com.umc.mada.user.dto;
-
-public class UserResponseDto {
-
-}
