+++ src/main/java/com/umc/mada/user/dto/user/UserRequestDto.java
@@ -1,4 +1,4 @@
-package com.umc.mada.user.dto;
+package com.umc.mada.user.dto.user;
 
 import lombok.*;
