+++ src/main/java/com/umc/mada/user/dto/user/UserResponseDto.java
@@ -1,0 +1,5 @@
+package com.umc.mada.user.dto.user;
+
+public class UserResponseDto {
+
+}
